requirements.txt
fastapi
uvicorn
pymupdf
langchain
langchain-google-genai
pytesseract
python-dotenv
numpy
chromadb
streamlit
pandas
plotly
python-multipart
pillow
requests
reportlab
fpdf
jinja2
httpx
nest-asyncio